+++ src/scorer/scorer.ts
@@ -7,13 +7,12 @@ import ScorerCache from './scorer_cache';
 import Storage from "../Storage";
 import Toot from '../api/objects/toot';
 import { ageString } from '../helpers/time_helpers';
-import { average, batchMap, percentileSegments, sumValues } from "../helpers/collection_helpers";
+import { average, batchMap, percentileSegments, sumArray, sumValues } from "../helpers/collection_helpers";
 import { config } from '../config';
 import { DEFAULT_WEIGHTS } from "./weight_presets";
 import { traceLog } from '../helpers/log_helpers';
-import { MinMaxAvgScore, NonScoreWeightName, ScoreName, ScoresStats, StringNumberDict, TootScore, WeightInfo, WeightName, Weights } from "../types";
+import { MinMaxAvgScore, NonScoreWeightName, ScoreName, ScoresStats, StringNumberDict, TootScore, TootScores, WeightedScore, WeightInfo, WeightName, Weights } from "../types";
 
-const SCORE_DIGITS = 3;  // Number of digits to display in the alternate score
 const SCORE_MUTEX = new Mutex();
 const SCORE_PREFIX = "scoreToots()";
 
@@ -22,9 +21,6 @@ const TRENDING_WEIGHTS = [
     ScoreName.TRENDING_TAGS,
     ScoreName.TRENDING_TOOTS,
 ];
-
-type WeightedAndUnweightedScores = Record<string, number | StringNumberDict>;
-type AlternateScoreDict = Record<string, number | WeightedAndUnweightedScores>;
 
 
 export default abstract class Scorer {
@@ -56,7 +52,7 @@ export default abstract class Scorer {
             console.warn(`${this.logPrefix()} not ready, scoring 0...`);
             return 0;
         } else {
-            const existingScore = toot.scoreInfo.rawScores[this.name];
+            const existingScore = toot.getIndividualScore("raw", this.name);
             console.debug(`${this.logPrefix()} Not ready but toot already scored (existing score: ${existingScore})`);
             return existingScore;
         }
@@ -74,45 +70,12 @@ export default abstract class Scorer {
     //   Static class methods   //
     //////////////////////////////
 
-    // Return a scoreInfo dict in a different format for the GUI (raw & weighted scores grouped in a subdict)
-    static alternateScoreInfo(toot: Toot): AlternateScoreDict {
-        if (!toot.scoreInfo) return {};
-
-        return Object.entries(toot.scoreInfo).reduce(
-            (scoreDict, [key, value]) => {
-                if (key == "rawScores") {
-                    scoreDict["scores"] = Object.entries(value).reduce(
-                        (scoreDetails, [scoreKey, scoreValue]) => {
-                            const weightedScore = toot.scoreInfo!.weightedScores[scoreKey as WeightName];
-
-                            // Only add non-zero scores to the dict
-                            if (scoreValue != 0) {
-                                scoreDetails[scoreKey] = {
-                                    unweighted: formatScore(scoreValue),
-                                    weighted: formatScore(weightedScore),
-                                };
-                            }
-
-                            return scoreDetails;
-                        },
-                        {} as WeightedAndUnweightedScores
-                    );
-                } else if (key != "weightedScores") {
-                    scoreDict[key] = formatScore(value as number);
-                }
-
-                return scoreDict;
-            },
-            {} as AlternateScoreDict
-        );
-    }
-
     // Compute stats about the scores of a list of toots
     static computeScoreStats(toots: Toot[], numPercentiles: number): ScoresStats {
         return Object.values(ScoreName).reduce((stats, scoreName) => {
             stats[scoreName] = {
-                raw: this.scoreStats(toots, scoreName, "rawScores", numPercentiles),
-                weighted: this.scoreStats(toots, scoreName, "weightedScores", numPercentiles),
+                raw: this.scoreStats(toots, scoreName, "raw", numPercentiles),
+                weighted: this.scoreStats(toots, scoreName, "weighted", numPercentiles),
             };
 
             return stats;
@@ -172,51 +135,57 @@ export default abstract class Scorer {
         const outlierDampener = getWeight(NonScoreWeightName.OUTLIER_DAMPENER);
         const timeDecayWeight = getWeight(NonScoreWeightName.TIME_DECAY) / 10;  // Divide by 10 to make it more user friendly
         const trendingMultiplier = getWeight(NonScoreWeightName.TRENDING);
-        // Initialize variables
-        const rawScores = {} as StringNumberDict;
-        const weightedScores = {} as StringNumberDict;
         // Do scoring
-        const scores = await Promise.all(scorers.map((s) => s.score(toot)));
+        const rawestScores = await Promise.all(scorers.map((s) => s.score(toot)));
 
         // Compute a weighted score a toot based by multiplying the value of each numerical property
         // by the user's chosen weighting for that property (the one configured with the GUI sliders).
-        scorers.forEach((scorer, i) => {
-            const scoreValue = scores[i] || 0;
-            rawScores[scorer.name] = scoreValue;
-            weightedScores[scorer.name] = scoreValue * (userWeights[scorer.name] ?? 0);
+        const scores: TootScores = scorers.reduce(
+            (scoreDict, scorer, i) => {
+                const rawScore = rawestScores[i] || 0;
+                let weightedScore = rawScore * (userWeights[scorer.name] ?? 0);
 
-            // Apply the TRENDING modifier but only to toots that are not from followed accounts or tags
-            if (realToot.isTrending() && (!realToot.isFollowed() || TRENDING_WEIGHTS.includes(scorer.name))) {
-                weightedScores[scorer.name] *= trendingMultiplier;
-            }
-
-            // Outlier dampener of 2 means take the square root of the score, 3 means cube root, etc.
-            if (outlierDampener > 0) {
-                const scorerScore = weightedScores[scorer.name];
-
-                // Diversity scores are negative so we temporarily flip the sign to get the root
-                if (scorerScore >= 0) {
-                    weightedScores[scorer.name] = Math.pow(scorerScore, 1 / outlierDampener);
-                } else {
-                    weightedScores[scorer.name] = -1 * Math.pow(-1 * scorerScore, 1 / outlierDampener);
+                // Apply the TRENDING modifier but only to toots that are not from followed accounts or tags
+                if (realToot.isTrending() && (!realToot.isFollowed() || TRENDING_WEIGHTS.includes(scorer.name))) {
+                    weightedScore *= trendingMultiplier;
                 }
-            }
-        });
+
+                // Outlier dampener of 2 means take the square root of the score, 3 means cube root, etc.
+                // TODO: outlierDampener is always greater than 0...
+                if (outlierDampener > 0) {
+                    const scorerScore = weightedScore;
+
+                    // Diversity scores are negative so we temporarily flip the sign to get the root
+                    if (scorerScore >= 0) {
+                        weightedScore = Math.pow(scorerScore, 1 / outlierDampener);
+                    } else {
+                        weightedScore = -1 * Math.pow(-1 * scorerScore, 1 / outlierDampener);
+                    }
+                }
+
+                scoreDict[scorer.name] = {
+                    raw: rawScore,
+                    weighted: weightedScore,
+                }
+
+                return scoreDict;
+            },
+            {} as TootScores
+        );
 
         // Multiple weighted score by time decay penalty to get a final weightedScore
         const decayExponent = -1 * Math.pow(toot.ageInHours(), config.scoring.timeDecayExponent);
         const timeDecayMultiplier = Math.pow(timeDecayWeight + 1, decayExponent);
-        const weightedScore = this.sumScores(weightedScores);
+        const weightedScore = this.sumScores(scores, "weighted");
         const score = weightedScore * timeDecayMultiplier;
 
         // Preserve rawScores, timeDecayMultiplier, and weightedScores for debugging
         const scoreInfo = {
-            rawScore: this.sumScores(rawScores),
-            rawScores,
+            rawScore: this.sumScores(scores, "raw"),
             score,
+            scores,
             timeDecayMultiplier,
             trendingMultiplier,
-            weightedScores,
             weightedScore,
         } as TootScore;
 
@@ -228,10 +197,10 @@ export default abstract class Scorer {
     private static scoreStats(
         toots: Toot[],
         scoreName: ScoreName,
-        scoreType: "rawScores" | "weightedScores",
+        scoreType: keyof WeightedScore,
         numPercentiles: number
     ): MinMaxAvgScore[] {
-        const getScoreOfType = (t: Toot) => t.scoreInfo?.[scoreType]?.[scoreName] ?? 0;
+        const getScoreOfType = (t: Toot) => t.getIndividualScore(scoreType, scoreName);
 
         return percentileSegments(toots, getScoreOfType, numPercentiles).map((segment) => {
             const sectionScores = segment.map(getScoreOfType);
@@ -247,18 +216,7 @@ export default abstract class Scorer {
     }
 
     // Add 1 so that time decay multiplier works even with scorers giving 0s
-    private static sumScores(scores: StringNumberDict | Weights): number {
-        return 1 + sumValues(scores);
+    private static sumScores(scores: TootScores, scoreType: keyof WeightedScore): number {
+        return 1 + sumArray(Object.values(scores).map((s) => s[scoreType]));
     }
-};
-
-
-export function formatScore(score: number): number {
-    if (typeof score != "number") {
-        console.warn(`${SCORE_PREFIX} formatScore() called with non-number:`, score);
-        return score;
-    }
-
-    if (Math.abs(score) < Math.pow(10, -1 * SCORE_DIGITS)) return score;
-    return Number(score.toPrecision(SCORE_DIGITS));
 };